restructured javascript modules
