+++ js/lib/DOM.js
@@ -1,0 +1,3 @@
+export async function DOMContentLoaded() {
+    await new Promise(resolve => document.addEventListener("DOMContentLoaded", resolve));
+}
