+++ js/lib/data-artist.js
@@ -1,34 +1,3 @@
-const currentURL = new URL(location.href);
-
-function decodedValue(key) {
-    try {
-        return decodeURIComponent(currentURL.searchParams.get(key));
-    } catch(err) {
-        throw new Error(`An error occurred attempting to decode the querystring value for ${key}. ${err.message}`);
-    }
-}
-
-const combineExistingValuesWithURLSearchParamValue =
-    (existingValues, key) =>
-        ({
-            ...existingValues,
-            [key]: decodedValue(key)
-        });
-
-export const queryStringPatcher =
-    data =>
-        Array.from(currentURL.searchParams.keys())
-            .filter(key => key in data)
-            .reduce(combineExistingValuesWithURLSearchParamValue, data);
-
-const defaultResponseParser = async res => await res.text();
-
-export async function fetcher(url, parseAction = defaultResponseParser) {
-    const fetched = await fetch(url);
-    if (!fetched.ok) throw new Error(`${fetched.statusText} (${fetched.status})`);
-    return await parseAction(fetched);
-}
-
 function stringify(object) {
     try {
         return JSON.stringify(object);
@@ -69,7 +38,7 @@ function pathExtract(data, schema) {
         }), {});
 }
 
-export function dataArtist(schema) {
+export default function dataArtist(schema) {
     return function(data) {
         return pathExtract(data, schema);
     }
